authorization finished

+++ Server/Controllers/Book.js
@@ -3,6 +3,8 @@ const BookModel = require("../Models/Book");
 
 /* Add Book To DB */
 exports.addBooks = async (req, res) => {
+  if(!req.user.isAdmin)
+    return res.status(401).send({error: "Unauthorized action"});
   const { name, category, author, description, bookImage } = req.body;
   const book = new BookModel({
     name,
@@ -57,6 +59,8 @@ exports.getOneBook = async (req, res) => {
 
 /* Delete one Book From DB */
 exports.deleteBook = async (req, res) => {
+  if(!req.user.isAdmin)
+    return res.status(401).send({error: "Unauthorized action"});
   const bookId = req.params.id;
   try {
     const deletedState = await BookModel.findByIdAndDelete(bookId);
@@ -72,6 +76,8 @@ exports.deleteBook = async (req, res) => {
 
 /* Update one Book From DB */
 exports.editBook = async (req, res) => {
+  if(!req.user.isAdmin)
+    return res.status(401).send({error: "Unauthorized action"});
   const { id } = req.params;
   const newBookData = req.body;
